refactor: move mood picker data class to utils in domain.
