+++ domain/src/main/java/com/amsterdam/domain/useCase/home/GetMoviesByMoodUseCase.kt
@@ -1,6 +1,6 @@
 package com.amsterdam.domain.useCase.home
 
-import com.amsterdam.domain.models.Mood
+import com.amsterdam.domain.utils.Mood
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Movie
