+++ ui/src/main/java/com/amsterdam/ui/screens/home/HomeScreen.kt
@@ -42,7 +42,7 @@ import com.amsterdam.designsystem.components.snackBar.SnackBarManager
 import com.amsterdam.designsystem.theme.AflamiTheme
 import com.amsterdam.designsystem.theme.AppTheme
 import com.amsterdam.designsystem.utils.ThemeAndLocalePreviews
-import com.amsterdam.domain.models.Mood
+import com.amsterdam.domain.utils.Mood
 import com.amsterdam.entity.category.MovieGenre
 import com.amsterdam.ui.R
 import com.amsterdam.ui.application.LocalNavManager
