+++ viewModel/src/main/java/com/amsterdam/viewmodel/home/HomeInteractionListener.kt
@@ -1,6 +1,6 @@
 package com.amsterdam.viewmodel.home
 
-import com.amsterdam.domain.models.Mood
+import com.amsterdam.domain.utils.Mood
 import com.amsterdam.entity.category.MovieGenre
 import com.amsterdam.viewmodel.shared.uiStates.MediaType
